requirements.txt
gym==0.26.2
hydra-core==1.3.2
ipykernel==6.19.2
joblib==1.2.0
langchain==0.0.181
omegaconf==2.3.0
openai==0.27.7
pandas==1.5.3
python-dotenv==1.0.0
tiktoken==0.4.0
wikipedia==1.4.0
torch # ==1.9.0+cu111
transformers==4.30.2
einops==0.6.1
sentence_transformers
matplotlib
# gdown # for downloading files from google drive (webshop)